refactor(logic): replace the parser by a router
