+++ x/logic/interpreter/fs/fs.go
@@ -11,15 +11,19 @@ import (
 // It will hold a list of handler that can resolve file URI and return the corresponding binary file.
 type FileSystem struct {
 	ctx    goctx.Context
-	parser Parser
+	router Router
 }
 
 // New return a new FileSystem object that will handle all virtual file on the interpreter.
 // File can be provided from different sources like CosmWasm cw-storage smart contract.
 func New(ctx goctx.Context, handlers []URIHandler) FileSystem {
+	router := NewRouter()
+	for _, handler := range handlers {
+		router.RegisterHandler(handler)
+	}
 	return FileSystem{
 		ctx:    ctx,
-		parser: Parser{handlers},
+		router: router,
 	}
 }
 
@@ -40,7 +44,7 @@ func (f FileSystem) Open(name string) (fs.File, error) {
 // ReadFile read the entire file at the uri provided.
 // Parse all handler and return the first supported handler file response.
 func (f FileSystem) ReadFile(name string) ([]byte, error) {
-	return f.parser.Parse(f.ctx, name)
+	return f.router.Open(f.ctx, name)
 }
 
 type Object []byte
